IGAPP-636: Refactor landing
