+++ native/src/components/Heading.tsx
@@ -1,13 +1,10 @@
 import React, { ReactElement } from 'react'
 import styled from 'styled-components/native'
 
-import { ThemeType } from 'build-configs'
-
 import EastereggImage from './EastereggImage'
 
 type HeadingProps = {
   clearResourcesAndCache: () => void
-  theme: ThemeType
 }
 const Wrapper = styled.View`
   display: flex;
@@ -15,9 +12,9 @@ const Wrapper = styled.View`
   align-items: center;
 `
 
-const Heading = ({ clearResourcesAndCache, theme }: HeadingProps): ReactElement => (
+const Heading = ({ clearResourcesAndCache }: HeadingProps): ReactElement => (
   <Wrapper>
-    <EastereggImage clearResourcesAndCache={clearResourcesAndCache} theme={theme} />
+    <EastereggImage clearResourcesAndCache={clearResourcesAndCache} />
   </Wrapper>
 )
